feat: update /dashboard/club Search components
