+++ src/app/dashboard/club/announcement/faq/page.tsx
@@ -1,5 +1,5 @@
+import FAQSearch from "@/components/dashboard/club/announcement/molecules/FAQSearch";
 import FAQList from "@/components/dashboard/club/announcement/organisms/FAQList";
-import FAQSearch from "@/components/dashboard/club/announcement/organisms/FAQSearch";
 
 export default function FAQPage() {
   return (
