+++ src/app/dashboard/club/announcement/page.tsx
@@ -1,6 +1,6 @@
 import AnnouncementOverview from "@/components/dashboard/club/announcement/organisms/AnnouncementOverview";
-import AnnouncementSearch from "@/components/dashboard/club/announcement/organisms/AnnouncementSearch";
-import AnnouncementList from "@/components/dashboard/club/announcement/organisms/AnnouncementList";
+import AnnouncementSearch from "@/components/dashboard/club/announcement/molecules/AnnouncementSearch";
+import AnnouncementList from "@/components/dashboard/shared/organisms/AnnouncementList";
 
 export default function AnnouncementPage() {
   return (
