+++ src/components/dashboard/club/announcement/molecules/AnnouncementSearch.tsx
@@ -9,11 +9,31 @@ import {
 
 const filterList = [
   { name: "전체 보기", value: "all" },
-  { name: "입금", value: "deposit" },
-  { name: "출금", value: "withdrawal" },
+  { name: "주무부서 공지", value: "company" },
+  {
+    name: (
+      <>
+        <span className="font-poppins font-bold">{"omo"}</span>
+        {" 공지"}
+      </>
+    ),
+    value: "omo",
+  },
+  {
+    name: (
+      <>
+        <span className="font-poppins font-bold">{"omo "}</span>
+        <span className="font-cochin font-bold text-lg leading-6">
+          {"shop"}
+        </span>
+        {" 공지"}
+      </>
+    ),
+    value: "omo-shop",
+  },
 ];
 
-export default function TransactionSearch() {
+export default function AnnouncementSearch() {
   const [currentSearchValue, setCurrentSearchValue] =
     useState<SearchValueWithFilter>(initialSearchValueWithFilter);
